+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0 (2023-10-25)
+
+- Add new arg datetime_format to az_blobcontainer_list module to control last_modified date format
+- Change blob last_modified date default format to match to_datetime ansible filter format.
+- Fix some sanity checks
+
 ## v1.0.0 (2023-10-23)
 
 - Major release for publishing